MB-63322 skip dropped non-vector instances
bg:
if there is a drop triggered for non-vector instances while
vector instances are under training (when vector and non-vector
are building in the same batch) then we will drop the instance

but the handler post training still considers those instances
when updating internal book keeping. since these instances are
deleted, when accessing from indexInstMap it gives the default
value to IndexInst struct. When such a struct is passed to
ClusterMgr, it panics when trying to process partition container

exp:
* do not drop non-vector instances too if it is in batch with
    vector insts under training
* skip indexInsts not found in indexInstMap
* also log all instances completing training for easy debugging

invariant:
handleIndexTrainingDone should run on the indexer main loop

Change-Id: I667062b516dee86e0ac3188b90b1000d5fd36840
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>