MB-63322 dont block panic on clusterMgrAgent close
if there is a panic in clusterMgrAgent, it will process the close
before throwing panic. this close can be blocked forever if the
lifecycleMgr is processing a req which needs indexer too as the
clusterMgrAgent run loop has already exited on panic.

hence if the system is going to panic, we should rather let it
go through rather than waiting for close

Change-Id: I45103b504b5178a2650a5ffa69d4f5c7fd6ea8c4
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>