MB-61128: Re-use codebook during rollback to Zero: PlasmaSlice
When the rollbackt to zero is triggered for slice, reuse the vector
metadata present in the slice before. The metadata such as nlist,
codeSize, codebookPath and codebook in the plasma slice still remain
relevant.
After the plasmaSlice's resetStores the information shaBuf and
quantizedCodeBuf is reset. quantizedCodeBuf needs to be initialised
again

Change-Id: I4e4a35dded94165d9d260e8599bda7b271d9d96f