MB-63254 Avoid copy of rows in vector scan pipeline
This change allows avoiding copy of rows returned by the
storage iterator in the vector scan pipeline for limit pushdown cases.
Each scan worker maintains a local TopK heap and works on the references
to the key/value returned by the iterator.

Before iterator close, only the TopK rows get copied and sent down the pipeline.
This saves the cost of copying all the rows returned by the iterator.

A new finishCallback has been introduced for storage reader methods.
This allows the storage to callback when it is done sending all the
qualified rows but before closing the iterator. At this point, the scan
pipeline will copy the final TopK rows from the local heap.

Change-Id: Ib5be90758677c94f65674f446a9380c2c42cdc4d