MB-63034 Add workerpool to rerank methods
For each partition, number of contexts available are defined
based on the "readersPerPartition" variable which is derived
based on nprobes and parallel centroid scans.

When doing re-ranking, "readersPerPartition" number of workers
will be spawned as a worker pool so that each worker can work
with independent context in parallel to fetch value from bhive
index.

Change-Id: If621b74e7437e251ea7646c6ef568812ea505543