MB-63034 Copy recordId, partnId and cid in row.Copy() method
Add a new method copyForBhive() to prevent unnecessary copy
of fields for other scans.

As BHIVE indexes would need partnId, recordId, cid for re-ranking
purposes, these fields are added to the row structure. When making
a copy of the row, also make a copy of these fields.

For "cid", it is sufficient to make a shallow copy as this is a read
only field derived from the scan request spans and the memory is valid
as long as the scan is alive.

Change-Id: I814c75a0cffae662b96bd9091a94fe32f46db130