MB-63600 Consider all proxies of realInst after prune
After an index prune, indexer will send an addInstance request
to projector with the new set of partitions that projector should
process. When merge and prune happen for the same index, there
exists a race condition where the partn list added by merge
is overwritten by the partn list during prune.

This can lead to a case where the partition list of an instance
at projector is partial and index becomes corrupted.

This patch fixes the issue by considering all caught up proxies
after pruning when sending stream update to projector. This will
ensure that projector contains all the partition list irrespective
of the order of merge and prune.

Change-Id: I26fec862d3dbde60b99518fb916b1ac085f36add