MB-63688 Fix the realInstId check when removing tombstone instances
BP to 7.6.4

When removing instances with zero partitions, only tombstone instances
are expected to be cleaned up. Real instances are expected to be
present in topology.

The fix for MB-63007 does this but the comparision with realInstId alone
is not sufficient. If replica of same definition undergoes the codepath,
then another replica with "0" partitions will be removed from topology.

Instead, the fix is updated to compare the RealInstId == 0 which is more
accurate considering all replicas as real index will always have
realInstId == 0.

Change-Id: I265285206738b8073b52692108dfd27cd42bd3de
(cherry picked from commit c117e094c1080199ce3cd617b5296a737945169b)