MB-63664 Move computation of buildTs before updating index state
This patch is required due to a regression from
https://review.couchbase.org/c/indexing/+/210366

If the state of the index is updated and then any error happens
during computation of buildTs, then retry of index build will
not work as indexer thinks index build is already in progress.

The index will be in INITIAL state forever without even starting
streams. To avoid this issue, the computation of buildTs is moved
before updating stream and state of the index.

The computation of buildTs is done after checking training necessity
to prevent unnecessary calls to KV.

Change-Id: I82da8927cfdd2856505114283a10d697f804dcfc