MB-63694 Update instance error before assining to indexInstMap
As indexInstMap is copy-on-value, any updates to inst has to be
done before assigning inst to indexInstMap. Later updates will
not be present in the indexInstMap and this is causing the
error state to be wrongly represented in meta.

Hence, updating the error state in local inst before updating
indexInstMap.

Change-Id: Ic3243f9d8270712fc1318617b79beb64af0902ca