MB-63837: Error out Rebal if the training of an index fails during Rebal
During Rebalance, if a vector index is replica repaired with mode DCP
or if the Rebalance is a DCP rebalance, then at the destination node
the index will be created and training will start.
Currently if the index errors out at the training phase, no action is
taken. The errored out index is observed and its state does not progress
causing reblance hang.

fix: with the fix if any index is errored out during training, then the
transfer token will be errored out. Failing the rebalance since the
index could not be built on destination node.

Change-Id: Ibf15340a214df7b0146d2f1a9f79f9ee8685b857