MB-63839: Cleanup codebooks explicitly during local index data cleanup
the issue occurs when a previous rebalance is cancelled or fails when
a TransferToken is in ShardRecoverState (just transitioned from
RestoreState) but the Proxy instance has not been created (not added
to the metadata repo)
As part of ShardRestoreState codebook had been transferred to the
respective index folder. So due to the cancel/failure, the token is
picked for clean-up but index is not found in the metadata and the
cleanup for the index folders are skipped.

fix: Explicitly remove the codebook folder whenever the local cleanup
of index data is called during cleanup

Change-Id: I713b32160ba24a0469dd194a157cc29b0f5537b8