MB-61103 Push includeColumns to bhive storage
The include column value given to indexer by projector
will be pushed to bhive storage with this patch.

The quantizedCodeBuf which is used to compute the quantizedCode
for vectors is being resued to stitch the quantizedCode and
includeColumns together.

Currently, scan pipeline extracts only the quantized code for
processing. Later patches will add support for inline filtering
in scans.

Change-Id: I56dab17d0c6ac422be2bf8ef3ebbbf5263e7fc40