MB-63838 Use first non-nil initial node in shardProxy
Prior to the fix done for MB-63370, the initialNode of a shard
proxy is set based on the first non-nil initial node in the list
of grouped indexes. This has resulted in a case where if there
are indexes with valid initialAlternateShardIds on node n1 but
if the first non-nil initial node is n2, then n2 is used as
initial node for shard proxy and rebalance failed with shard
not found error as the initial node is incorrectly set for shard
transfer (as shard exists on n1 and not n2).

To fix this case, the first non-nil initial node with valid initial
alternate shardIds is set as initial node for shard proxy. However,
this has resulted in another regression - If none of the indexes
in the list have valid initial alternate shard proxy, the proxy
will have nil initial node. Planner thinks that this shard is being
repaired and will skip filtering incase of cyclic movements.

E.g., consider a replica being present on n2, n4, n5 and if planner
decides following movements n2 -> n1, n1 -> n5, n4 -> n2 for a replica,
then the movement n2 -> n1 will be considered repair on n1 and
similarly n1 -> n5 will be considered repair on n5 by planner due
to initialNode being nil for shard proxies.

Hence, filter solution is skipped in both cases while both nodes
contain the replicas. This leads to rebalance failure with "Index
already exists" error.

To avoid this case, planner falls back to setting the first non-nil
initialNode as initialNode of shard proxy. However, this only happens
after planner ensures that there is no index with valid initial
alternate shardIds.

Change-Id: Ic0c61c283668b7676390b0b86d5cb2f713d3ee86