MB-61124: Populate numCentroids in NumberOfCentroids
- Propagate numCentroids when the index training phase transitions
to TRAINING_COMPLETED, as it is finalized during training.
- Ensure topology changes are observed by the query client;
numCentroids is updated in local metadata within the index instance
definition, making it available for secondary_index.
- Set NumberOfCentroids at index definition creation if nlist is
specified in vector metadata; otherwise, update NumberOfPartitions
with the correct value after training completion.

Situations tested:
- Failover node containing subset of partitions, NumPartitions returned is same as definition.
- Restart query service with one indexer node failed over, NumPartitions returned is same as definition.

Caution:
Number of instances will be 0 if following cases:
1. Index definition has been created no instance has been created yet
2. If all index instances have been deleted.
This will cause NumPartitions returned to be 0, though this should be expected as no instance is present.

Change-Id: I6740798f6035d68ae6964e8128dba2a6070e1cfd