MB-64216 Add AllowRerank() option in secondaryIndex6 interface
This will allow query to read if reranking is enabled in index
defition and will handle the rerank option during scan time
appropriately

Note: This is just a stub function for query team to make any
changes. The implementation of this function would change in
the later set of patches i.e. after implementing the option to
disable storing full vectors in bhive indexes.

Change-Id: Ib66257769aafb3d69596fd567849be7165f4d012