MB-64276 Track rebalance transfer phase per partition
Otherwise, if multiple partitions are moving across multiple
tokens and if one token finishes movement and the other does not,
it is possible that the rebalance transfer phase is updated
incorrectly.

Any external event like collection drop will lead to a case
where the instance will be dropped while transfer is in progress
and this can lead to rebalance failure.

Change-Id: Id9d55f88b9fa748cafc778f2ddca81f4cbe44a1f