MB-63638 Use minSeqno. from all storage snapshots during rollback
possible that the timestamps for each vbucket are different
for different indexes.

Consider one such timestamp combination for 2 vbuckets.
Index i1 has {10, 20} as timestamp and
Index i2 has {15, 18} as timestamp

When indexer receives a rollback and rollback timestamp has to
be computed for these 2 indexes, current code would do either
of the following:

a. {10, 20}.AsRecentTs({15, 18}) - This will evaluate to false
   as 18 < 20 and indexer will restart the stream from {15, 18}.
   This means that all mutations between 10 and 15 will be lost
   for vb: 0

b. {15, 18}.AsRecentTs({10, 20}) - This will evaluate to false
   as 10 < 15 and indexer will restart stream from {10, 20}.
   This means that all mutations between 18 and 19 will be lost
   for vb: 1

In either cases there is a data loss due to the usage of AsRecentTs()
method when computing the timestamp. To avoid the data loss, indexer
will now do ComputeMinTs() which will compute the minimum seqno.
across all available seqnos. and use that for restarting the stream.

For the timestamps given in above example, the restart timestamp
would be ComputeMinTs({15, 18}, {10, 20}) which will result in
a new timestamp of {10, 18}. This will ensure that no mutations
are missed when restarting the streams.

Change-Id: I6794388289288b651b19caaac7f0e0d8937cda53