MB-63641 Send projector initiated streamEnds only after DequeueMutations() exits
Otherwise, dfCloseStream command will send streamEnd to downstream
and clears the vbstreams book-keeping. As the stream is deleted,
any mutations that are in mutation queue will be considered spurious.

To avoid that case, projector will first terminate mutation queue
before sending streamEnds so that genServer() will take care of
sending streamEnds and the mutations will not be considered spurious.

Change-Id: I3559039b4b71d1934f464ae6d8267c345b4d760a