MB-61122 Add support for partition sets
Current implementation for partition elimination is built on two
invariants:

1. All partition keys are included in the secondary expressions.
2. All partition keys have equality predicates.

For BHIVE indexes, invariant (1) is not possible as a BHIVE index
contains only one expression which is a vector field. To eliminate
scan on unnecessary partitions, query has introduced the concept of
partition sets.

E.g., consider the below index
create vector index idx_1 on default(sift vector)
       include(name, age, email, zipcode, county, country)
       partition by hash(age, gender)

"age" and "gender" are not a part of the secondary expressions.

Consider the below scan:

select count(*) from default where name is not null and
       age = 25 and gender = “male”

Since there is equality predicate on "age" and "gender", there
is only one partition this request needs to scan. However, due
to lack of partition elimination, all partitions will be scanned.

This issue is adderssed by partition sets where query will send
a list of all values with equality/IN predicate as a partition set
to GSI client. GSI client can compute the hash of these values
and filter unnecessary partitions.

In this example, query will send the partition set (25, "male) to
GSI client. GSI client will compute hash(25, "male") and the scan
will be routed only to that partition.

Change-Id: I3e5b08bd602dc044501ede933db79416ab69b0e2