MB-61106 Support for null/missing for non-leading vectors
Projector will follow the below rules for indexing NULL/MISSING
entries

a. If VECTOR entry is MISSING/NULL and it is leading key, the
   document will not be indexed
b. If VECTOR entry is MISSING/NULL and it is non-leading key,
   we index MISSING/NULL for the VECTOR entry
c. If the VECTOR entry is an invalid and it is leading key,
   the document will not be indexed
d. If the VECTOR entry is invalid and it is non-leading key,
   the VECTOR entry will be indexed as NULL

Change-Id: I31edb54680196632d0c48caa917cd9aaa8922a32