MB-64378 Inline filter pushdown to BHIVE
This patch will push down a callback to BHIVE storage engine.
The callback will be set at the time of initializing new
iterator.

The callback will take the decoded "meta" value, extract the
"includeColumn" value and then evaluate the includeColumn value
agains the expression in the scan request.

If the expression evaluates to true, the callback will return "true".
Otherwise, it will return false and the row scan be skipped from
further processing.

Change-Id: I7fefd8ef8c363a5d7128fe99621169ffde7fbe87