MB-65012 Disable "nodes" with alter index replica_count
BP of MB-64982 to 7.6.6

When changing replica count (specifically incrementing replica
count), if the required alternateID of shard exists on a node
that is outside the list of nodes specified in the "nodes" clause
with alter index command, then indexer goes ahead and creates the
index on the list of nodes specified.

This can lead to a case where same alternate shardID can exist on
more than one node in the cluster. This breaks the fundamental
invariant of file based rebalance - There should be a strict one-one
mapping between alternate shardID and nodes. Other than rebalance times,
an alternate shardID should exist only on one node in the cluster.

This patch disables specifying "nodes" clause with alter index command
when the action is "replica_count"

Change-Id: Iad5d5340b03c057aab5d531a6ecac6042eba3f52