MB-64098: Introduce Shard type identifier during file based Rebalance
Introduced new constant to identify the type of shard undergoing the
movement during file based rebalance.
This is essential because the RPC function for plasma and Bhive which
perform the actual shard related operation are different. With the shard
type identifier the shardIds within the transfer token can be identified

Change-Id: I1e1c16b186a89760c45b6511ae67c7510e4440c7