MB-64098: Populate the Shard types from Transfer token on Source and Dest Node
The patch introduces the flow to enable populating the shard type for
each of the Shard undergoing movement.
It will be populated during the first processing stage of the Transfer
token on both Source and Destination, when the nodes start processing
the tokens respectively, the Shard Rebalancer will send a message to
StorageMgr (then to shard transfer manager) to store the shard type.

This will be done during ShardTokenCreated state on source and
ShardTokenScheduledOnSource state on destination

Change-Id: Id976be7740fc5523e86a0a1d7fa054ff51e64465