MB-64098: Update the Transfer shard flow to work for Bhive shards
Using the shard type mapper, the shard storage manager identifies which
type of shard is being transferred to the destination and calls the
respective library call for the transfer.
If the shard type is unset, it is treated like an error, where the error
is populate in the Transfer token. This is done as the only case where
the shard type is UNSET is when two different index types are present
in the same the Transfer token OR when a ShardId collision between
plasma and bhive has taken place

Change-Id: I0c22786d8e109dc72ee2acd7849a1fd22d7743fc