MB-64975: Fix race: Token in commit state to initiate others in batch
race: There can be a race condition where a transfer token in
TransferTokenAccept state is waiting for other token in the batch to
transition from TransferTokenCreated -> TransferTokenAccept but the
token in TransferTokenCreated can move to TransferTokenCommit instead,
making the token in TransferTokenAccept never progressing causing
DCP Rebalance hang.

fix: To tackle this move to initiate state only if all the tokens have
been accepted by the destination as well as there is no other token in
committed state yet to be processed. If there is a token in
TransferTokenCommit let the last token in batch handle transition of
accepted tokens.

Change-Id: I6461404466a507145e2d8517a02c6d42cdedf2d9