MB-64379 Build missing bhive graph during warmup
If any ACTIVE bhive index has metadata status of the graph as not ready,
then build the graph again. This can happen if indexer crashed after
index moved to ACTIVE state but before graph could be built or its
status could be updated in the metadata. The BuildGraph storage API will
ignore the request if graph is already built.

Change-Id: I25bedc98def82888cbb1ccf46a1f8a20a35fa84d