MB-64098: Populate the shard types during the cleanup phase for TT
The shard types are populated at the time of transfer token cleanup as
well. This is due to the fact that the node could have restarted during
the rebalance, with the restart/failover the shard type information will
not be available.
The shard types are repopualted for the cases of no restarts

Change-Id: If2bd45fa9dde6d5a3a0219ac7dafb12f94342b57