[BP 7.6.7] MB-65377: Defer closing the RPC server after the transfer token cleanup
During cleanup of transfer token (failed rebalance or successful) calls
to RPC Server are made. If due to Rebalance cancel being called if the
RPC server is closed before the cleanup is triggered then residual files
can remain on the node.

The patch aims at deferring the closure of RPC server for rebalances
from the Cancel call to after the cleanup of tokens have gone through
The closure of RPC is now done as part of OnRebalanceDoneLocked, this
function is called explicitly for Cancel and eventually for all
Rebalances and Failover.
Condition have been added to Rebalance Service Manager to check if the
closure of RPC server should be called (non-serverless mode and Shard
affinity can be achieved).
* For failovers, a DCP rebalancer is used - RPC server closure won't be
called
* For successful rebalances, same behaviour as before.
* For failed rebalances, RPC server closed afterwards the cleanup called
from done callback.
* For cancelled rebalances, the RPC server is closed after the cleanup
has been executed called directly from OnRebalanceDoneLocked.

Change-Id: If8b34fd0f8c6a692eb1f63d67261a9244e473b61