MB-65203 Cache annotated values and exprContext in inline filtering
Otherwise, every iteration of inline filter callback will create
a new map and an annotated value which will eventually be garbage
collected.

Since the key's to the map are fixed for a scan, it is ok to cache
these values to prevent excessive allocations.

Also, remove 2 different implementations for filtering include columns.
Have only one implementation in inlineFilterCb

Change-Id: Iac29225ae1ca8a279fd63bc66ab0eeff12173fcb