MB-65215 force refetch of localMeta on creating new metadata
with doing async reads of local meta in shard rebalancer, it
can happen that destination creates new instances but local
meta in shard rebalancer is stale.

if this happens for a new partition being added to an already
existing instance (real instance) then the stale meta leads
to shard rebalancer reading real instance information and it
believes that the new partition to create is already active.
while this is happening, the instance could actually be still
under recovery and hence this breaks Shard Rebalances.

the instance will come out of recovery but it wont get merged
to the real instance as the rebalance is already over. the
instance will remain in RebalPending state due to this.

future attempts to rebuild this partition in rebalances will
continue to fail with overlapping partitions error as the
partition exists but its not in the correct state

this patch fixes the above by forcing refresh on local meta
if we are creating new metadata

Change-Id: Ief98ef0ec048a51af8d04f1d3ece50f3b5899c07
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>