[BP 7.2.8] MB-65375: Set min and max iteration per temp for Replica Repair scenario
BP MB-65347 to 7.2.8

If the max and min iteration is not set, it causes the the replica
repair to just proceed with the initial placement of the new replica.
During the initial placement, only replica violation is checked.
Server group are automatically handled due to priority of nodes but the
initial placement can hit into equivalent index check.
So when the parameter is not set the new indexes are not moved around
and no neighbor for solution are ever found

With the fix, we are setting the param for min and max iteration, so
even during replica repair a solution's neighbors are found and so that
indexes are correctly placed

Change-Id: Ibb7de2997108b190ff7c77de8e526679533156b5