MB-65065 add vector related classification in IndexSpecs
currently, we did not populate the IndexSpecs with the
new IsBhive and IsCompositeVector index fields. this
leads to different category indexes sharing the same
shard when they are created by planner in the create
index DDL statements (not in rebalances).

fix the separation issue by correctly adding the same
to specs and then using to create more appropriate
IndexUsage

Change-Id: Ie88a174635ff9d646df3bb993b1a4507e446d222
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>