MB-65421: UseShardId as specified by Rebalancer and populate ShardIds
If rebalancer provides ShardId at the time of initializing, these
ShardIds are used so that reset stores will map the bhive instances to
the shard on which the instance is transferred.

RebalRunning flag is also set to ensure that UseShardId is only
populated and used during Rebalance. Just like plasma, the flag is
cleared either when shards are unlocked or at the end of rebalance.

Change-Id: I784073032390edb1e6227f794e48b407a0ba20cd