MB-65138 Integrate timestampedCounts stats request with storage manager
Any request to the REST server on stats_manager for timestamped
counts will go to storage manager. Storage manager will respond with
all the timestamped counts request.

Following HTTP codes are returned by this method:
a) 200 - For successful response (or) FDB storage mode
b) 503 - If index service is in bootstrap
c) 500 - If any error is observed while processing this request
d) 400 - If request is not GET request
e) 408 - If request is timed out

Additionally, a URL parameter - doLog is added for debuggability.

Change-Id: I6acf8ac063b9abe0108a4fca35e61fb5e8f1e926