MB-65138 Add stats for corrupted indexes
Two new stats: numCorruptedIndexes and corruptedIndexMap
are added at stats manager.

These stats are set by indexer::monitorItemsCount when
indexer finds some corrupted indexes. If there are any
corrupt indexes, the num_corrupt_indexes is incremented
and corruptedIndexesMap is updated with the index that
is corrupt.

If the index corruption is resolved or if the node looses
ownership of items_count mismatch computation during rebalance,
it will reset the counters. During rebalance, other nodes would
update the stats.

Change-Id: Ic7843fc4470d96d67899c5978710cf8433002019