MB-65138 Add is_corrupted stat to prometheusMetricsHigh endpoint
If there are any corrupted indexes in the list, it will be added
to prometheusMetricsHigh endpoint for end-users to consume.

The behaviour will be as follows:
a) Index process started at t = 0. num_corrupted_indexes = 0 and no
   indexes in corrupted list till t = 10. The endpoint will not
   publish any list to prometheus

b) At t = 10, there is one index found to be corrupt. Let us call this
   index: "i1". So, num_corrupt_indexes is incremented and
   "i1:is_corrput": true is published to prometheus.

c) At t = 20, the corruption on index "i1" is resolved by some means.
   At this point, indexer will stop publishing "i1:is_corrupt" to prometheus
   i.e. the stat will be absent in the output of prometheus endpoint.

Change-Id: Ia0ef7467904fbf213c8bb9b5c2fb347158ef54ad