MB-65203 Pass include column spans to indexer
When possible, query would push the spans on include columns
to indexer. Indexer can use this spans to filter rows without
the need for decoding the rows (As decoding the rows is a very
expensive operation which takes more CPU and memory).

This patch adds necessary changes to pass the spans from query
process to GSI client to index service.

Change-Id: I0c2da98cd121de478ede9cd22967a8ac63766217