MB-65203 Filter based on include column spans when available
When spans are available for include columns, then use those
spans for filtering include column expressions.

A scan row is considered to be qualified if the row matches
atleast one scan. Only if none of the scans match, then
the row is considered to be disqualified from further
processing.

In some cases where the number of scans are large,
query may not push down the spans to indexer. In such
cases, indexer will fallback to evaluating the row
via inline filter expression.

Change-Id: I1c2d040ecd44eb4a7e5e3121ab33521e52f1c4b2