MB-65037: Make the recoverRebalance cleanup condition same as setCleanupPending
When setting the cleanupPending flag the condition of either rebalance
running or rebalance token to be present is checked. But after the
Indexer becomes ready, the recoverRebalance code will cleanup the local
rebalance token. For cleanup only the rebalance running flag is checked
This can leave out the cleanup of the rebalance token is the flag is set
to false, but the token is not nil.

This particular condition can hit when after the failed/cancelled
rebalance, a failover is triggered. A successful failover will cleanup
the rebalance running flag but not rebalance token

Change-Id: I35d68be9b39f18a5f05340fe2f7946813c2aa4fb