MB-65203 Generate n1qlVal based on code instead of tmp buffer
As tmp buffer can get overridden with data. When there are more
than one string entries in code, then the first and second
entries can potentially share the same data due to the way the
value is constructed. This can lead to incorrect results.

Instead, the value is generated based on "code" as code does
not change throughout the processing of row.

Change-Id: I73c15388babb0008aa0e8368fbac2efb3c732a73