MB-65290: Consider only the participating dest nodes in smart batching
Currently, all the keep nodes are considered when dividing the batchSize
during smart batching. This is incorrect for cases when the exclude
nodes value is set for some nodes. With the exclude node value some
nodes will not participate in rebalance, and hence the batchSize will
get divided by the num of Keep nodes instead of the nodes participating
in the rebalance.
An example of such case is Swap Rebalance, where only the outgoing and
incoming node participate in rebalance (unless redistribute indexes flag
is turned on).
There are cases where some replica repair might be happening during swap
rebalance, this cases is also handled as the numIndexNodes is not static
it is dependent on the TTs remaining to be built. So if the BatchSize
gets divided based on the TTs that are going to be built in future

Change-Id: I6cf0b73522e238df072174e3d91ae017b3032a9c