MB-65470 Reduce vb concurrency for sampling
For range scan, each vb has a fixed timeout of 180 seconds. If too many
vb range scans are started together and indexer is slow in processing
the sampled vectors, it can lead to a timeout at KV.

This patch reduces the concurrency to avoid the issue. Tested this patch
for perf regression with 100M 128 dimension test. The time spent on
sampling KV data is similar with the changes.

Also, increased the buffer channel size to allow more queueing at the
indexer in case it is slow to process sampled vectors.

Change-Id: I3bd2c8649d5a6e838116d4edf687da10b5a4e740