MB-65631: Cap scaled Jemalloc Overhead and Indexer heap for minMemory
Currently. the overhead memory calculation blindly scales the overhead
memory coming form the difference of mem_used - mem_used_storage.
These stats are point in time data meaning that for transient workload
the jemalloc fragmentation can shoot up and the planner will use the
large difference between memory_total_storage - memory_used_storage to
blindly scale up based on the resident ratio of the index.

The fix proposes capping the contribution of jemalloc fragmentation to
the set maxFragThreshold value. This ensures that the minMemory
calculation is inline with our sizing recommendation to account for
jemalloc fragmentation.

Change-Id: Icfd983375370f2289d55e1978bccf917ea58a998