MB-65417 only consider slot for re-use when index aligns
bg: shard dealer tries to reuse slots which are present on nodes
where index is placed. but it does not bind the replica ID of
index and shard. it expects that an index can move around the nodes.
that is only valid for index creation requests and hence during mixed
mode rebalances, we can have rebalance failures when rebalance ends up
attempting to move around these indexes but thatt cannot happen as
index replicas already exist (and hence index creation with new replic
ID fails)

exp: only move around instances if we know that this movement is
possible. this is done on each replica level in the implementation.
if there is a replica repair, we might consider lost replica eligible
for movement but other replicas will fail to move and hence that slot
will be skipped.

Change-Id: I97e8b361f91e7904695f5122005d1f567459384b
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>