MB-65987 Comapare items count mismatch only if bucketSeqnos > indexSeqnos
During bucket flush, if one indexer node is able to process
the flush and other indexer node is yet to process, then a
false alert will be raised that the index is corrupted.

Hence, added a check that bucket seqnos must be greater
than or equal to index seqnos. for this check to happen.

Change-Id: I07f47dac0071ae14a31aece1eb0b2bc328f8e0c4