MB-66155: Fix panic in NeedsReplan when finding the index
NeedsReplan can be triggered when the same indexSlot can exist on
multiple nodes. i.e. it does not mean that if we find the indexSlot on
one Node, the index itself will be present on it
For e.g.
	Node1: slotA-r0 { idx1-r0, idx2-r0 }
	Node2: slotA-r0 { idx3-r0, idx4-r0 }
if we are replanning for idx3-r0, the slotMap will have an entry for
Node1 as well but idx3-r0 will not be present here.
the value offset == -1 represents that, Hence the patch focuses on
skipping the removal or addition of this index from the node

Change-Id: Id9629547d65f03bda5bb860909f60beb4f074864