MB-66218 Do not delete index inst if recovery fails
If recovery fails due to plasma error, indexer will update
the meta info and delete the index from book-keeping. Any
further attempts from lifecycle manager to delete the indexer
will fail as the index will not be a part of the indexInstMap.

This will lead to a case where other objects like stats are
not properly updated and this can lead to stale results. This
patch addresses the issue by skipping the delete. Actual deletion
will happen when lifecycle manager attempts to delete the index
as a part of token cleanup.

Change-Id: Iaf44bb23aca277e32cf8d7776910388d8b0e528f