MB-64809 use cluster comptability to compute node version
bg: if a new node has spun up and there is a node in the cluster
which is not reachable, then it is possible that the version info
for it may not be available. this info is also available as a part
of the cluster comptability which can be used to compute major and
minor version

exp: if the version string validation fails, then fallback to calculate
version from cluster comptability. if that too fails and the node is
unhealthy, then we ignore such nodes from cluster info cache validation

assumption: version string is a volatile info (lost on system restart)
but clusterCompatibility is persistent info

pool endpoint sample struct considered for this patch -
/pools/default
    - []nodes
	* status (healthy, unhealthy)
	* clusterMembership (active, inactive, inactiveAdded, etc)
	* nodeUUID
	* clusterCompatibility (int)
	* version (eg 8.0.0-2855-enterprise)
	...

Change-Id: Ie9cbc0d684e0c2a2e3b8d4f9da4eb7196facabf1
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>