MB-66288 Use ComputeDistanceEncoded for PQ cosine
ComputeDistanceEncoded is used for L2 for both PQ and SQ.
For COSINE, it is only used for PQ as there is recall drop
for SQ.

Also set large buffer size for PQ + Cosine.

This change is not done for BHIVE as it doesn't use
ComputeDistanceEncoded for PQ as majority of the perf gains come
from using a large batch size which doesn't work well for BHIVE design.

Change-Id: I8ce1d62b30fe334fb3659dba8d5536064cb83090