MB-66266 allow slice closure for drop of duplicate index
bg: if we have a shard which is undergoing movement and that
shard has a duplicate index then the drop for such duplicate
index will happen but slice closuer will get skipped until
end of rebalance. since this is a duplicate index, it's
metadata is not a part of the rebalance transfer tokens
because it is to be dropped before rebalance has started.
hence slice closuer for this index gets skipped (during
rebalance inititalisation) and later it never gets closed.
this leads to rebalance hang as the source node won't move
transfer tokens to the next step until this slice gets closed

exp:
since we are going to drop the index and it is to happen
before rebalance starts, we can close the slice as closure
won't race with shard transfer. so we can allow drop of such
index.

Change-Id: I484c29ea0ea6fe689eb14fd7fb5e953044cfb14b
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>